requirements.txt
redis
numpy
pandas
pyyaml
pymysql
requests
retrying
pycryptodome
python-dateutil

fastapi[all]==0.95.1
captcha==0.4                               # 验证码